requirements/prod.txt
# Everything needed in production

# Flask
Flask==0.10.1
MarkupSafe==0.23
Werkzeug==0.10.4
Jinja2==2.7.3
itsdangerous==0.24

# Data
Flask-Misaka==0.3.0
pyyaml>=3.11

# Bootstrap
Flask-Bootstrap==3.3.5.2

# Forms
Flask-WTF==0.11
WTForms==2.0.2

# Deployment
gunicorn>=19.1.1

# Assets
Flask-Assets==0.10
cssmin>=0.2.0
jsmin>=2.0.11

# Caching
Flask-Cache>=0.13.1

# Debug toolbar
Flask-DebugToolbar==0.9.2
